start conversion of touch story
